// x-pack/platform/plugins/shared/onechat/public/application/components/tools/index_search/form/types/index_search_tool_form_types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

export interface OnechatIndexSearchToolFormData {
  name: string;
  description: string;
  pattern: string;
  tags: string[];
}